fixes in data type and router

+++ models/time.js
@@ -1,7 +1,7 @@
 'use strict';
 module.exports = function (sequelize, DataTypes) {
   const Time = sequelize.define('Time', {
-    current_time: DataTypes.FLOAT
+    current_time: DataTypes.INTEGER
   /*, {
     classMethods: {
       associate: function(models) {
